provider/ec2: remove StartInstance volume handling
EBS volumes will only be provisioned via the dynamic
storage provisioner now. We'll continue to allocate
all instance stores, so the EBS provisioner can later
use them if desired.
