+++ provider/ec2/ebs.go
@@ -11,6 +11,7 @@ import (
 	"github.com/juju/utils/set"
 	"gopkg.in/amz.v3/ec2"
 
+	"github.com/juju/juju/constraints"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/storage"
 	"github.com/juju/juju/storage/poolmanager"
@@ -43,6 +44,30 @@ const (
 const (
 	volumeInUse        = "VolumeInUse"
 	attachmentNotFound = "InvalidAttachment.NotFound"
+)
+
+const (
+	// minRootDiskSizeMiB is the minimum/default size (in mebibytes) for ec2 root disks.
+	minRootDiskSizeMiB uint64 = 8 * 1024
+
+	// provisionedIopsvolumeSizeMinGiB is the minimum disk size (in gibibytes)
+	// for provisioned IOPS EBS volumes.
+	provisionedIopsvolumeSizeMinGiB = 10 // 10 GiB
+
+	// volumeSizeMaxGiB is the maximum disk size (in gibibytes) for EBS volumes.
+	volumeSizeMaxGiB = 1024 // 1024 GiB
+
+	// maxProvisionedIopsSizeRatio is the maximum allowed ratio of IOPS to
+	// size (in GiB), for provisioend IOPS volumes.
+	maxProvisionedIopsSizeRatio = 30
+
+	// devicePrefix is the prefix for device names specified when creating volumes.
+	devicePrefix = "/dev/sd"
+
+	// renamedDevicePrefix is the prefix for device names after they have
+	// been renamed. This should replace "devicePrefix" in the device name
+	// when recording the block device info in state.
+	renamedDevicePrefix = "xvd"
 )
 
 func init() {
@@ -184,6 +209,11 @@ func (v *ebsVolumeSource) CreateVolumes(params []storage.VolumeParams) (vols []s
 			}
 		}
 	}()
+
+	// TODO(axw) if preference is to use ephemeral, use ephemeral
+	// until the instance stores run out. We'll need to know how
+	// many there are and how big each one is. We also need to
+	// unmap ephemeral0 in cloud-init.
 
 	// First, validate the params before we use them.
 	for _, p := range params {
@@ -403,4 +433,98 @@ func (v *ebsVolumeSource) DetachVolumes(attachParams []storage.VolumeAttachmentP
 		}
 	}
 	return nil
+}
+
+var errTooManyVolumes = errors.New("too many EBS volumes to attach")
+
+// blockDeviceNamer returns a function that cycles through block device names.
+//
+// The returned function returns the device name that should be used in
+// requests to the EC2 API, and and also the (kernel) device name as it
+// will appear on the machine.
+//
+// See http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/block-device-mapping-concepts.html
+func blockDeviceNamer(numbers bool) func() (requestName, actualName string, err error) {
+	const (
+		// deviceLetterMin is the first letter to use for EBS block device names.
+		deviceLetterMin = 'f'
+		// deviceLetterMax is the last letter to use for EBS block device names.
+		deviceLetterMax = 'p'
+		// deviceNumMax is the maximum value for trailing numbers on block device name.
+		deviceNumMax = 6
+	)
+	var n int
+	letterRepeats := 1
+	if numbers {
+		letterRepeats = deviceNumMax
+	}
+	return func() (string, string, error) {
+		letter := deviceLetterMin + (n / letterRepeats)
+		if letter > deviceLetterMax {
+			return "", "", errTooManyVolumes
+		}
+		deviceName := devicePrefix + string(letter)
+		if numbers {
+			deviceName += string('1' + (n % deviceNumMax))
+		}
+		n++
+		realDeviceName := renamedDevicePrefix + deviceName[len(devicePrefix):]
+		return deviceName, realDeviceName, nil
+	}
+}
+
+// getBlockDeviceMappings translates constraints into BlockDeviceMappings.
+//
+// The first entry is always the root disk mapping, followed by instance
+// stores (ephemeral disks).
+func getBlockDeviceMappings(cons constraints.Value) ([]ec2.BlockDeviceMapping, error) {
+	rootDiskSizeMiB := minRootDiskSizeMiB
+	if cons.RootDisk != nil {
+		if *cons.RootDisk >= minRootDiskSizeMiB {
+			rootDiskSizeMiB = *cons.RootDisk
+		} else {
+			logger.Infof(
+				"Ignoring root-disk constraint of %dM because it is smaller than the EC2 image size of %dM",
+				*cons.RootDisk,
+				minRootDiskSizeMiB,
+			)
+		}
+	}
+
+	// The first block device is for the root disk.
+	blockDeviceMappings := []ec2.BlockDeviceMapping{{
+		DeviceName: "/dev/sda1",
+		VolumeSize: int64(mibToGib(rootDiskSizeMiB)),
+	}}
+
+	// Not all machines have this many instance stores.
+	// Instances will be started with as many of the
+	// instance stores as they can support.
+	blockDeviceMappings = append(blockDeviceMappings, []ec2.BlockDeviceMapping{{
+		VirtualName: "ephemeral0",
+		DeviceName:  "/dev/sdb",
+	}, {
+		VirtualName: "ephemeral1",
+		DeviceName:  "/dev/sdc",
+	}, {
+		VirtualName: "ephemeral2",
+		DeviceName:  "/dev/sdd",
+	}, {
+		VirtualName: "ephemeral3",
+		DeviceName:  "/dev/sde",
+	}}...)
+
+	return blockDeviceMappings, nil
+}
+
+// mibToGib converts mebibytes to gibibytes.
+// AWS expects GiB, we work in MiB; round up
+// to nearest GiB.
+func mibToGib(m uint64) uint64 {
+	return (m + 1023) / 1024
+}
+
+// gibToMib converts gibibytes to mebibytes.
+func gibToMib(g uint64) uint64 {
+	return g * 1024
 }
