+++ provider/ec2/ebs_test.go
@@ -5,6 +5,7 @@ package ec2_test
 
 import (
 	"sort"
+	"strconv"
 
 	"github.com/juju/errors"
 	"github.com/juju/names"
@@ -13,6 +14,7 @@ import (
 	"gopkg.in/amz.v3/ec2/ec2test"
 	gc "gopkg.in/check.v1"
 
+	"github.com/juju/juju/constraints"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/environs/jujutest"
 	"github.com/juju/juju/instance"
@@ -378,4 +380,88 @@ func (s *ebsVolumeSuite) TestDetachVolumes(c *gc.C) {
 	// Test idempotent
 	err = vs.DetachVolumes(params)
 	c.Assert(err, jc.ErrorIsNil)
+}
+
+type blockDeviceMappingSuite struct {
+	testing.BaseSuite
+}
+
+var _ = gc.Suite(&blockDeviceMappingSuite{})
+
+func (*blockDeviceMappingSuite) TestBlockDeviceNamer(c *gc.C) {
+	var nextName func() (string, string, error)
+	expect := func(expectRequest, expectActual string) {
+		request, actual, err := nextName()
+		c.Assert(err, jc.ErrorIsNil)
+		c.Assert(request, gc.Equals, expectRequest)
+		c.Assert(actual, gc.Equals, expectActual)
+	}
+	expectN := func(expectRequest, expectActual string) {
+		for i := 1; i <= 6; i++ {
+			request, actual, err := nextName()
+			c.Assert(err, jc.ErrorIsNil)
+			c.Assert(request, gc.Equals, expectRequest+strconv.Itoa(i))
+			c.Assert(actual, gc.Equals, expectActual+strconv.Itoa(i))
+		}
+	}
+	expectErr := func(expectErr string) {
+		_, _, err := nextName()
+		c.Assert(err, gc.ErrorMatches, expectErr)
+	}
+
+	// First without numbers.
+	nextName = ec2.BlockDeviceNamer(false)
+	expect("/dev/sdf", "xvdf")
+	expect("/dev/sdg", "xvdg")
+	expect("/dev/sdh", "xvdh")
+	expect("/dev/sdi", "xvdi")
+	expect("/dev/sdj", "xvdj")
+	expect("/dev/sdk", "xvdk")
+	expect("/dev/sdl", "xvdl")
+	expect("/dev/sdm", "xvdm")
+	expect("/dev/sdn", "xvdn")
+	expect("/dev/sdo", "xvdo")
+	expect("/dev/sdp", "xvdp")
+	expectErr("too many EBS volumes to attach")
+
+	// Now with numbers.
+	nextName = ec2.BlockDeviceNamer(true)
+	expect("/dev/sdf1", "xvdf1")
+	expect("/dev/sdf2", "xvdf2")
+	expect("/dev/sdf3", "xvdf3")
+	expect("/dev/sdf4", "xvdf4")
+	expect("/dev/sdf5", "xvdf5")
+	expect("/dev/sdf6", "xvdf6")
+	expectN("/dev/sdg", "xvdg")
+	expectN("/dev/sdh", "xvdh")
+	expectN("/dev/sdi", "xvdi")
+	expectN("/dev/sdj", "xvdj")
+	expectN("/dev/sdk", "xvdk")
+	expectN("/dev/sdl", "xvdl")
+	expectN("/dev/sdm", "xvdm")
+	expectN("/dev/sdn", "xvdn")
+	expectN("/dev/sdo", "xvdo")
+	expectN("/dev/sdp", "xvdp")
+	expectErr("too many EBS volumes to attach")
+}
+
+func (*blockDeviceMappingSuite) TestGetBlockDeviceMappings(c *gc.C) {
+	mapping, err := ec2.GetBlockDeviceMappings(constraints.Value{})
+	c.Assert(err, jc.ErrorIsNil)
+	c.Assert(mapping, gc.DeepEquals, []awsec2.BlockDeviceMapping{{
+		VolumeSize: 8,
+		DeviceName: "/dev/sda1",
+	}, {
+		VirtualName: "ephemeral0",
+		DeviceName:  "/dev/sdb",
+	}, {
+		VirtualName: "ephemeral1",
+		DeviceName:  "/dev/sdc",
+	}, {
+		VirtualName: "ephemeral2",
+		DeviceName:  "/dev/sdd",
+	}, {
+		VirtualName: "ephemeral3",
+		DeviceName:  "/dev/sde",
+	}})
 }
