+++ provider/ec2/environ.go
@@ -499,9 +499,7 @@ func (e *environ) StartInstance(args environs.StartInstanceParams) (*environs.St
 	}
 	var instResp *ec2.RunInstancesResp
 
-	blockDeviceMappings, volumes, volumeAttachments, err := getBlockDeviceMappings(
-		*spec.InstanceType.VirtType, &args,
-	)
+	blockDeviceMappings, err := getBlockDeviceMappings(args.Constraints)
 	if err != nil {
 		return nil, errors.Annotate(err, "cannot create block device mappings")
 	}
@@ -540,13 +538,6 @@ func (e *environ) StartInstance(args environs.StartInstanceParams) (*environs.St
 	// TODO(axw) tag all resources (instances and volumes), for accounting
 	// and identification.
 
-	if err := assignVolumeIds(inst, volumes, volumeAttachments); err != nil {
-		if err := e.StopInstances(inst.Id()); err != nil {
-			logger.Errorf("failed to stop instance: %v", err)
-		}
-		return nil, err
-	}
-
 	if multiwatcher.AnyJobNeedsState(args.MachineConfig.Jobs...) {
 		if err := common.AddStateInstance(e.Storage(), inst.Id()); err != nil {
 			logger.Errorf("could not record instance in provider-state: %v", err)
@@ -563,10 +554,8 @@ func (e *environ) StartInstance(args environs.StartInstanceParams) (*environs.St
 		AvailabilityZone: &inst.Instance.AvailZone,
 	}
 	return &environs.StartInstanceResult{
-		Instance:          inst,
-		Hardware:          &hc,
-		Volumes:           volumes,
-		VolumeAttachments: volumeAttachments,
+		Instance: inst,
+		Hardware: &hc,
 	}, nil
 }
